recipes/natural_language_processing/rag/app/requirements.txt
langchain-openai==0.1.7
langchain==0.1.20
chromadb==0.5.23
sentence-transformers==2.7.0
streamlit
protobuf
pypdf==4.2.0
pymilvus>=2.5.0
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation
